Extract dashboard grab ops to separate module to satisfy source file length limit
